Avoided a conversion warning

+++ modules/juce_audio_formats/codecs/juce_WavAudioFormat.cpp
@@ -664,7 +664,8 @@ namespace WavFileHelpers
                         {
                             MemoryBlock mb;
                             input.readIntoMemoryBlock (mb, (ssize_t) infoLength);
-                            values.set (types[i], String::createStringFromData ((const char*) mb.getData(), mb.getSize()));
+                            values.set (types[i], String::createStringFromData ((const char*) mb.getData(),
+                                                                                (int) mb.getSize()));
                             break;
                         }
                     }
